Now uses Rectangles.

+++ src/IPositionable.java
@@ -1,6 +1,6 @@
 public interface IPositionable {
-    int getX();
-    int getY();
-    int getWidth();
-    int getHeight();
+    double getX();
+    double getY();
+    double getWidth();
+    double getHeight();
 }
